The next iteration of Hub. Requires the hubssolib Gem. Stores user login details in a secure cookie rather than the session, opening up the potential for cross-application access. Still need to move session expiry and last-recorded-URL stuff into a cookie, otherwise it's all done. Many bugs fixed, though some may have been introduced in splitting functionality into the Gem.
No solution presently to whether or not the User object will be understood
or accessible in any way from other applications (not yet tried). Checking
this version in because it's reached a stage of reasonable stability,
before the next step of attempting wider integration.

Remember, this won't work at all unless hubssolib 0.0.3 is installed.